kit and network separation
